:truck: rename class
