+++ src/main/resources/application.properties
@@ -4,6 +4,7 @@ server.port=8080
 server.shutdown=graceful
 spring.lifecycle.timeout-per-shutdown-phase=30s
 
+#Eureka ??
 spring.application.name=g2nee-shop
 eureka.client.fetch-registry=true
 eureka.client.register-with-eureka=true
@@ -11,13 +12,14 @@ eureka.instance.prefer-ip-address =true
 eureka.client.service-url.defaultZone=http://g2nee:g2nee@133.186.150.129:8761/eureka
 #eureka.client.registry-fetch-interval-seconds=5
 
+# DB ??
 spring.jpa.hibernate.naming.implicit-strategy=org.hibernate.boot.model.naming.ImplicitNamingStrategyLegacyJpaImpl
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.show-sql=true
 spring.jpa.open-in-view=false
 
+# SecureKey ??
 nhncloud.url=https://api-keymanager.nhncloudservice.com/
 nhncloud.path=keymanager/v1.0/appkey/
 nhncloud.appKey=${APP_KEY}
